installer/requirements.txt
PyQt6>=6.6.1
pywin32>=306
wmi>=1.5.1
psutil>=5.9.6
pillow>=10.1.0
pyusb>=1.2.1
pyinstaller>=6.3.0
requests>=2.31.0
pyyaml>=6.0.1
humanize>=4.9.0
